client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "prometheus-client>=0.24.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
